Revert "Add Caduq point, spline and patch"
