+++ Caduq/src/SandboxApp.cpp
@@ -1,5 +1,5 @@
 #include <Vizir.h>
-//#include "Vizir/Core/EntryPoint.h"
+#include "Vizir/Core/EntryPoint.h"
 
 #include "Vizir/Platform/OpenGL/OpenGLShader.h"
 #include "Vizir/Renderer/Shader.h"
@@ -19,7 +19,6 @@
 #include <glm/gtc/type_ptr.hpp>
 #include "glm/ext.hpp" // TEMP
 
-// #include "SandboxCq.h"
 #include "SandboxSplines.h"
 
 
@@ -231,24 +230,24 @@ private:
 	glm::vec3 m_SquareColor = { 0.2f, 0.1f, 0.8f };
 };
 
-// class GameApplication : public Vizir::Application
-// {
-// public:
-// 	GameApplication()
-// 	{
-// 		PushLayer(new SandboxSplines());
-//
-// 		/*m_CameraLayer = new CameraLayer(&m_Camera);
-// 		PushLayer(m_CameraLayer);*/
-// 	}
-// 	~GameApplication()
-// 	{
-//
-// 	}
-// 	private:
-// };
-//
-// Vizir::Application* Vizir::CreateApplication()
-// {
-// 	return new GameApplication();
-// }
+class GameApplication : public Vizir::Application
+{
+public:
+	GameApplication()
+	{
+		PushLayer(new SandboxSplines());
+
+		/*m_CameraLayer = new CameraLayer(&m_Camera);
+		PushLayer(m_CameraLayer);*/
+	}
+	~GameApplication()
+	{
+
+	}
+	private:
+};
+
+Vizir::Application* Vizir::CreateApplication()
+{
+	return new GameApplication();
+}
