+++ Geotrup/src/Geometry/Point.cpp
@@ -2,16 +2,10 @@
 
 #include "Point.h"
 
-#include "Eigen/Core"
 namespace Geometry
 {
     Point::Point(double x, double y, double z)
         : m_x{ x }, m_y{ y }, m_z{ z }
-    {
-    }
-
-    Point::Point(Eigen::Vector3d pos)
-        : m_x{ pos(0) }, m_y{ pos(1) }, m_z{ pos(2) }
     {
     }
