+++ Geotrup/src/Geometry/Point.h
@@ -1,5 +1,5 @@
-#ifndef GO_POINT_H
-#define GO_POINT_H
+#ifndef POINT_H
+#define POINT_H
 
 #include <Eigen/Dense>
 
@@ -15,7 +15,6 @@ namespace Geometry
         public:
         Point() = default;
         Point(double x, double y, double z = 0.0);
-        Point(Eigen::Vector3d pos);
 
         void SetPosition(double x, double y, double z = 0.0);
         Eigen::Vector3d GetPosition();
