+++ Geotrup/src/Geometry/Spline.h
@@ -1,13 +1,20 @@
-#ifndef GO_SPLINE_H
-#define GO_SPLINE_H
+#ifndef SPLINE_H
+#define SPLINE_H
 
 #include <Eigen/Dense>
 #include <tuple>
 
-#include "Geo.h"
+#include "Point.h"
 
 namespace Geometry
 {
+    struct SplinePoint
+    {
+        Geometry::Point point{};
+        Eigen::RowVector3d tangent{ 0.0, 0.0, 0.0 };
+        double tension{ 0 };
+    };
+
     class Spline
     {
         private:
@@ -21,7 +28,6 @@ namespace Geometry
         Spline(SplinePoint startPoint, SplinePoint endPoint);
         Eigen::MatrixXd Mesh(Eigen::ArrayXd u, const int MESH_SIZE);
         std::tuple<Eigen::MatrixXd, Eigen::VectorXi> GetFemMesh();
-        Geometry::Mesh GetGfxMesh();
         double GetLength();
     };
 }
